+++ WorkManager/pom.xml
@@ -1,7 +1,7 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
 	<modelVersion>4.0.0</modelVersion>
-	<groupId>test.test.ojt</groupId>
+	<groupId>jp.kigami.ojt</groupId>
 	<artifactId>WorkManager</artifactId>
 	<packaging>war</packaging>
 	<version>0.0.1-SNAPSHOT</version>
