+++ WorkManager/src/main/java/jp/kigami/ojt/common/util/ValidationResult.java
@@ -1,0 +1,24 @@
+package jp.kigami.ojt.common.util;
+
+public class ValidationResult {
+
+	private String errorMsg;
+
+	private boolean checkResult;
+
+	public String getErrorMsg() {
+		return errorMsg;
+	}
+
+	public void setErrorMsg(String errorMsg) {
+		this.errorMsg = errorMsg;
+	}
+
+	public boolean isCheckResult() {
+		return checkResult;
+	}
+
+	public void setCheckResult(boolean checkResult) {
+		this.checkResult = checkResult;
+	}
+}
