--- WorkManager/src/main/java/test/test/ojt/common/util/MsgCodeDef.java
@@ -1,12 +1,0 @@
-package test.test.ojt.common.util;
-
-public class MsgCodeDef {
-
-	public static final String CONFIRM_ERROR = "e.validation.001";
-
-	public static final String SIZE_ERROR = "e.validation.002";
-
-	private MsgCodeDef() {
-	}
-
-}
