--- WorkManager/src/main/java/test/test/ojt/model/User.java
@@ -1,49 +1,0 @@
-package test.test.ojt.model;
-
-import java.io.Serializable;
-
-public class User implements Serializable {
-
-	private static final long serialVersionUID = 1L;
-
-	private int userId;
-
-	private String userName;
-
-	private String password;
-
-	private String confirmPassword;
-
-	public int getUserId() {
-		return userId;
-	}
-
-	public void setUserId(int userId) {
-		this.userId = userId;
-	}
-
-	public String getUserName() {
-		return userName;
-	}
-
-	public void setUserName(String userName) {
-		this.userName = userName;
-	}
-
-	public String getPassword() {
-		return password;
-	}
-
-	public void setPassword(String password) {
-		this.password = password;
-	}
-
-	public String getConfirmPassword() {
-		return confirmPassword;
-	}
-
-	public void setConfirmPassword(String confirmPassword) {
-		this.confirmPassword = confirmPassword;
-	}
-
-}
