--- WorkManager/src/main/java/test/test/ojt/servlet/MenuServlet.java
@@ -1,38 +1,0 @@
-package test.test.ojt.servlet;
-
-import java.io.IOException;
-
-import javax.servlet.RequestDispatcher;
-import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import javax.servlet.http.HttpSession;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-@WebServlet("/Menu")
-public class MenuServlet extends HttpServlet {
-
-	private static final long serialVersionUID = 1L;
-
-	private static final Logger logger = LoggerFactory.getLogger(MenuServlet.class);
-
-	@Override
-	public void doGet(HttpServletRequest request, HttpServletResponse response) {
-
-		// 検索条件削除
-		HttpSession session = request.getSession();
-		session.removeAttribute("criteria");
-		logger.info("検索条件を削除。");
-
-		RequestDispatcher dispatcher = request.getRequestDispatcher("/WEB-INF/jsp/menu.jsp");
-		try {
-			dispatcher.forward(request, response);
-		} catch (ServletException | IOException e) {
-			logger.error("フォワード失敗", e);
-		}
-	}
-}
