+++ WorkManager/src/main/webapp/WEB-INF/jsp/user/userRegistConfirm.jsp
@@ -1,4 +1,4 @@
-<%@page import="test.test.ojt.model.User"%>
+<%@page import="jp.kigami.ojt.model.User"%>
 <%@ page language="java" contentType="text/html; charset=UTF-8"
 	pageEncoding="UTF-8"%>
 <!DOCTYPE html>
